+++ sentry-android-core/src/test/java/io/sentry/android/core/ActivityLifecycleIntegrationTest.kt
@@ -370,6 +370,8 @@ class ActivityLifecycleIntegrationTest {
             check {
                 assertEquals(SpanStatus.OK, it.status)
             },
+            anyOrNull(),
+            anyOrNull(),
             anyOrNull()
         )
     }
@@ -391,6 +393,8 @@ class ActivityLifecycleIntegrationTest {
             check {
                 assertEquals(SpanStatus.UNKNOWN_ERROR, it.status)
             },
+            anyOrNull(),
+            anyOrNull(),
             anyOrNull()
         )
     }
@@ -406,7 +410,7 @@ class ActivityLifecycleIntegrationTest {
         sut.onActivityCreated(activity, fixture.bundle)
         sut.onActivityPostResumed(activity)
 
-        verify(fixture.hub, never()).captureTransaction(any(), anyOrNull())
+        verify(fixture.hub, never()).captureTransaction(any(), anyOrNull(), anyOrNull())
     }
 
     @Test
@@ -417,7 +421,7 @@ class ActivityLifecycleIntegrationTest {
         val activity = mock<Activity>()
         sut.onActivityPostResumed(activity)
 
-        verify(fixture.hub, never()).captureTransaction(any(), anyOrNull())
+        verify(fixture.hub, never()).captureTransaction(any(), anyOrNull(), anyOrNull())
     }
 
     @Test
@@ -430,7 +434,7 @@ class ActivityLifecycleIntegrationTest {
         sut.onActivityCreated(activity, fixture.bundle)
         sut.onActivityDestroyed(activity)
 
-        verify(fixture.hub).captureTransaction(any(), anyOrNull())
+        verify(fixture.hub).captureTransaction(any(), anyOrNull(), anyOrNull(), anyOrNull())
     }
 
     @Test
@@ -499,7 +503,7 @@ class ActivityLifecycleIntegrationTest {
         sut.onActivityCreated(mock(), mock())
 
         sut.onActivityCreated(mock(), fixture.bundle)
-        verify(fixture.hub).captureTransaction(any(), anyOrNull())
+        verify(fixture.hub).captureTransaction(any(), anyOrNull(), anyOrNull(), anyOrNull())
     }
 
     @Test
@@ -512,7 +516,7 @@ class ActivityLifecycleIntegrationTest {
         sut.onActivityCreated(activity, mock())
         sut.onActivityResumed(activity)
 
-        verify(fixture.hub, never()).captureTransaction(any(), any())
+        verify(fixture.hub, never()).captureTransaction(any(), any(), anyOrNull())
     }
 
     @Test
@@ -539,7 +543,7 @@ class ActivityLifecycleIntegrationTest {
         sut.onActivityCreated(activity, mock())
         sut.onActivityResumed(activity)
 
-        verify(fixture.hub).captureTransaction(any(), anyOrNull())
+        verify(fixture.hub).captureTransaction(any(), anyOrNull(), anyOrNull(), anyOrNull())
     }
 
     @Test