+++ sentry/src/main/java/io/sentry/NoOpHub.java
@@ -120,7 +120,8 @@ public final class NoOpHub implements IHub {
   public @NotNull SentryId captureTransaction(
       final @NotNull SentryTransaction transaction,
       final @Nullable TraceState traceState,
-      final @Nullable Object hint) {
+      final @Nullable Object hint,
+      final @Nullable ProfilingTraceData profilingTraceData) {
     return SentryId.EMPTY_ID;
   }