+++ sentry/src/test/java/io/sentry/SentryTracerTest.kt
@@ -8,11 +8,14 @@ import com.nhaarman.mockitokotlin2.never
 import com.nhaarman.mockitokotlin2.spy
 import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
+import io.sentry.protocol.App
+import io.sentry.protocol.Request
 import io.sentry.protocol.User
 import java.util.Date
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
+import kotlin.test.assertNotEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertNull
 import kotlin.test.assertSame
@@ -118,6 +121,8 @@ class SentryTracerTest {
             check {
                 assertEquals(it.transaction, tracer.name)
             },
+            anyOrNull(),
+            anyOrNull(),
             anyOrNull()
         )
     }
@@ -153,6 +158,8 @@ class SentryTracerTest {
                     assertEquals(emptyMap(), it.tags)
                 }
             },
+            anyOrNull(),
+            anyOrNull(),
             anyOrNull()
         )
     }
@@ -169,6 +176,8 @@ class SentryTracerTest {
                 assertEquals(1, it.spans.size)
                 assertEquals("op1", it.spans.first().op)
             },
+            anyOrNull(),
+            anyOrNull(),
             anyOrNull()
         )
     }
@@ -300,6 +309,8 @@ class SentryTracerTest {
                     assertEquals(SpanStatus.OK, it.status)
                 }
             },
+            anyOrNull(),
+            anyOrNull(),
             anyOrNull()
         )
 
@@ -342,7 +353,7 @@ class SentryTracerTest {
         val transaction = fixture.getSut(waitForChildren = true)
         transaction.startChild("op")
         transaction.finish()
-        verify(fixture.hub, never()).captureTransaction(any(), any())
+        verify(fixture.hub, never()).captureTransaction(any(), any(), anyOrNull(), anyOrNull())
     }
 
     @Test
@@ -351,7 +362,7 @@ class SentryTracerTest {
         val child = transaction.startChild("op")
         child.finish()
         transaction.finish()
-        verify(fixture.hub).captureTransaction(any(), anyOrNull())
+        verify(fixture.hub).captureTransaction(any(), anyOrNull(), anyOrNull(), anyOrNull())
     }
 
     @Test
@@ -372,7 +383,7 @@ class SentryTracerTest {
         val transaction = fixture.getSut(waitForChildren = true)
         val child = transaction.startChild("op")
         child.finish()
-        verify(fixture.hub, never()).captureTransaction(any(), any())
+        verify(fixture.hub, never()).captureTransaction(any(), any(), anyOrNull(), anyOrNull())
     }
 
     @Test
@@ -380,12 +391,14 @@ class SentryTracerTest {
         val transaction = fixture.getSut(waitForChildren = true)
         val child = transaction.startChild("op")
         transaction.finish(SpanStatus.INVALID_ARGUMENT)
-        verify(fixture.hub, never()).captureTransaction(any(), any())
+        verify(fixture.hub, never()).captureTransaction(any(), any(), anyOrNull(), anyOrNull())
         child.finish()
         verify(fixture.hub, times(1)).captureTransaction(
             check {
                 assertEquals(SpanStatus.INVALID_ARGUMENT, it.status)
             },
+            anyOrNull(),
+            anyOrNull(),
             anyOrNull()
         )
     }
@@ -404,6 +417,8 @@ class SentryTracerTest {
                 assertEquals(SpanStatus.DEADLINE_EXCEEDED, it.spans[0].status)
                 assertEquals(SpanStatus.DEADLINE_EXCEEDED, it.spans[1].status)
             },
+            anyOrNull(),
+            anyOrNull(),
             anyOrNull()
         )
     }
@@ -472,6 +487,8 @@ class SentryTracerTest {
             check {
                 assertEquals("val", it.getExtra("key"))
             },
+            anyOrNull(),
+            anyOrNull(),
             anyOrNull()
         )
     }
@@ -489,6 +506,8 @@ class SentryTracerTest {
                     assertEquals("val", it["key"])
                 }
             },
+            anyOrNull(),
+            anyOrNull(),
             anyOrNull()
         )
     }